Replace Welcome with ChatWindow chatbot component
Replaces the Welcome component with a new ChatWindow component, including new styles and logic for a chat-based interface. Updates the main page to use ChatWindow, removes all Welcome-related files, and renames the test file accordingly. Updates app metadata to reflect the new chatbot focus.
