+++ app/page.tsx
@@ -1,9 +1,15 @@
-import { Welcome } from '../components/Welcome/Welcome';
+'use client';
+import { AppShell } from '@mantine/core';
+import { ChatWindow } from '../components/ChatWindow/ChatWindow';
 
 export default function HomePage() {
   return (
     <>
-      <Welcome />
+      <AppShell padding="md">
+        <AppShell.Main>
+          <ChatWindow />
+        </AppShell.Main>
+      </AppShell>
     </>
   );
 }
