+++ components/ChatWindow/ChatWindow.module.css
@@ -1,0 +1,111 @@
+.chatWindow {
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  width: 400px;
+  height: 500px;
+  border: 1px solid #e0e0e0;
+  border-radius: 12px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  background: white;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  z-index: 1000;
+}
+
+.minimizedChat {
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  z-index: 1000;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  transition: transform 0.2s ease;
+}
+
+.minimizedChat:hover {
+  transform: scale(1.1);
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 16px 20px;
+  border-bottom: 1px solid #e0e0e0;
+  background: #f8f9fa;
+  flex-shrink: 0;
+}
+
+.messagesArea {
+  flex: 1;
+  padding: 16px 20px;
+  overflow-y: auto;
+}
+
+.messagesContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.message {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  padding: 8px 12px;
+  border-radius: 8px;
+  max-width: 80%;
+}
+
+.message.user {
+  align-self: flex-end;
+  background: #007bff;
+  color: white;
+}
+
+.message.assistant {
+  align-self: flex-start;
+  background: #f1f3f4;
+  color: #333;
+}
+
+.roleLabel {
+  font-size: 11px;
+  opacity: 0.8;
+  text-transform: capitalize;
+}
+
+.messageText {
+  word-wrap: break-word;
+}
+
+.errorMessage {
+  text-align: center;
+  padding: 8px;
+  background: #fee;
+  border-radius: 4px;
+  margin: 8px 0;
+}
+
+.inputArea {
+  padding: 16px 20px;
+  border-top: 1px solid #e0e0e0;
+  background: #f8f9fa;
+  flex-shrink: 0;
+}
+
+.messageInput {
+  flex: 1;
+}
+
+.messageInput textarea {
+  border: 1px solid #ddd;
+  resize: none;
+  font-family: inherit;
+  line-height: 1.4;
+}
+
+.messageInput textarea:focus {
+  border-color: #007bff;
+}
