+++ components/ChatWindow/ChatWindow.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, userEvent } from '@/test-utils';
-import { Welcome } from './Welcome';
+import { ChatWindow } from './ChatWindow';
 
 // Mock the ClientRateLimiter
 jest.mock('../../app/lib/utils/api-helpers', () => ({
@@ -17,27 +17,21 @@ describe('Welcome component', () => {
     jest.clearAllMocks();
   });
 
-  it('renders the welcome title', () => {
-    render(<Welcome />);
-    expect(screen.getByText(/Welcome to your/)).toBeInTheDocument();
-    expect(screen.getByText('Starter')).toBeInTheDocument();
-  });
-
   it('renders input field and buttons', () => {
-    render(<Welcome />);
+    render(<ChatWindow />);
     expect(screen.getByLabelText('Ask a Question')).toBeInTheDocument();
     expect(screen.getByText('Ask Question')).toBeInTheDocument();
     expect(screen.getByText('Reset')).toBeInTheDocument();
   });
 
   it('displays remaining requests count', () => {
-    render(<Welcome />);
+    render(<ChatWindow />);
     expect(screen.getByText(/You have \d+ questions remaining/)).toBeInTheDocument();
   });
 
   it('allows user to type in input field', async () => {
     const user = userEvent.setup();
-    render(<Welcome />);
+    render(<ChatWindow />);
 
     const input = screen.getByLabelText('Ask a Question');
     await user.type(input, 'Hello world');
@@ -47,7 +41,7 @@ describe('Welcome component', () => {
 
   it('shows error when trying to submit empty input', async () => {
     const user = userEvent.setup();
-    render(<Welcome />);
+    render(<ChatWindow />);
 
     const submitButton = screen.getByText('Ask Question');
     await user.click(submitButton);
@@ -57,7 +51,7 @@ describe('Welcome component', () => {
 
   it('resets form when reset button is clicked', async () => {
     const user = userEvent.setup();
-    render(<Welcome />);
+    render(<ChatWindow />);
 
     const input = screen.getByLabelText('Ask a Question');
     const resetButton = screen.getByText('Reset');
