--- components/Welcome/Welcome.story.tsx
@@ -1,7 +1,0 @@
-import { Welcome } from './Welcome';
-
-export default {
-  title: 'Welcome',
-};
-
-export const Usage = () => <Welcome />;
